+++ sdk/@launchdarkly/observability/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @launchdarkly/observability
 
+## [0.4.4](https://github.com/launchdarkly/observability-sdk/compare/observability-0.4.3...observability-0.4.4) (2025-10-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.22.1
+
+## [0.4.3](https://github.com/launchdarkly/observability-sdk/compare/observability-0.4.2...observability-0.4.3) (2025-10-01)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.22.0
+
+## [0.4.2](https://github.com/launchdarkly/observability-sdk/compare/observability-0.4.1...observability-0.4.2) (2025-09-24)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.21.1
+
 ## [0.4.1](https://github.com/launchdarkly/observability-sdk/compare/observability-0.4.0...observability-0.4.1) (2025-09-09)